requirements.txt
geopandas
rapidfuzz